Implement the ability to suspend and resume a ring
The consumer signals that it doesn't want any more items i.e.
the ring should be considered suspended. The producer notices
the request, acknowledges it and complies.

When the consumer wants more items again, it clears the suspend
signal. The producer notices the request and complies.

Note that the producer must now poll the state regularly with the
'state' function in order to see the requests.

Signed-off-by: David Scott <dave.scott@citrix.com>