Journal: add a 'wake_up' bool as a latch
Previously we used the 'wake_up_time' and compared it with 'now' to
figure out whether we should flush or not. Instead we set 'wake_up'
to true every time the counter expires, so if we do

  Alarm.reset 0
  Alarm.reset 5

The alarm will trigger twice. Note also the countdown thread sets
t.thread to None before signalling that the countdown has expired,
so the reset function will know whether to set up another timer or
not.

Signed-off-by: David Scott <dave.scott@citrix.com>