CA-179407: make it safe to call `*.state` functions in parallel threads
Each ring instance had a shared scratch buffer for reading and writing
sectors. If one thread called `push` while another thread calls `state`
then the following can happen:

- `push` reads a sector from the data portion of the ring
- `state` read the consumer pointer *into the same buffer*
- `push` writes a modified sector to the data portion of the ring

The result is the data portion of the ring has a copy of either the
producer or the consumer pointer.

This patch makes it safe by always allocating a private buffer. The
buffer is only 1 sector in size and the cost of doing this is dwarfed
by the cost of performing the I/O.

Signed-off-by: David Scott <dave.scott@citrix.com>