test: Test interleaving ring suspend and advance
Currently there is a race condition between these operations with the caching
of the ring states. If Consumer.advance starts and calls set_consumer to
advance the consumer pointer on it's sector it will then block. Then another
thread may call suspend which will call set_consumer to set the suspend bit
which will also block on IO. In some cases the write to the suspend bit will be
clobbered by the write to advance the pointer.

This can be seen by applying the below patch to the advance code which causes
the new test to consistently fail.

diff --git a/lib/ring.ml b/lib/ring.ml
index cc80121..c2edc6b 100644
--- a/lib/ring.ml
+++ b/lib/ring.ml
@@ -519,6 +519,9 @@ module Consumer = struct
         let open C in
         let sector = alloc t.info.B.sector_size in
         let consumer' = { t.consumer with consumer = consumer } in
+        let open Lwt in
+        Lwt_unix.sleep 0.1 >>= fun () ->
+        let open C in
         set_consumer ~queue:t.queue ~client:t.client t.disk sector consumer' >>= fun () ->
         t.consumer <- consumer';
         return (`Ok ())

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>