Fix deadlock, and retry failed journalled ops
There was a deadlock when shutting down the journal, as the
condition wait was not checking for the request to shutdown.

Additionally, this commit also ensures that the background
thread is created even if the initial run of the journal
entries fails, and that if there is a failure, the journal
periodically attempts to retry.

Signed-off-by: Jon Ludlam <jonathan.ludlam@citrix.com>